src-server/src-server/pyproject.toml
[project]
name = "pop-server"
version = "0.1.0"
description = "Path of Purpose — Backend API server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyjwt>=2.9",
    "stripe>=11.0",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pop_server*"]
